pytorch_support/requirements.txt
absl-py==0.4.1
astor==0.7.1
backports.weakref==1.0.post1
certifi==2018.8.24
chardet==3.0.4
enum34==1.1.6
flatbuffers==1.9
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
graphpipe==1.0.4
grpcio==1.14.2
idna==2.6
Markdown==2.6.11
mock==2.0.0
numpy==1.14.5
onnx==1.3.0
onnx-tf==1.1.2
pbr==4.2.0
Pillow==5.2.0
protobuf==3.6.1
requests==2.18.4
six==1.11.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
typing==3.6.6
typing-extensions==3.6.5
urllib3==1.22
Werkzeug==0.14.1
